fixed tests + clarified reader exceptions

+++ src/main/java/ca/ulaval/glo4003/spamdul/infrastructure/ui/infractions/dto/InfractionResponse.java
@@ -1,7 +1,8 @@
 package ca.ulaval.glo4003.spamdul.infrastructure.ui.infractions.dto;
 
 public class InfractionResponse {
+
   public String code;
   public String reason;
-  public int amount;
+  public double amount;
 }
